make redeem logic runnable in separate binary
swapd can now be run with independent parts. It could be split into
running a chain sync process, a redeem process, a preimage process and a
public server process. As long as all processes use the same database,
they can work independently.

Recommended is to split swapd into two parts:
- A server process, using --no-redeem --no-preimage and --no-chain
- A background process, isolated from the internet, using --no-servers

This way the important things like chain sync, redeeming and checking
for preimages are not vulnerable to DDOS attack.