add lnd support
Add LND support to swapd. The lightning client can now be either a cln
client or lnd client. It is determined whether an LND or CLN client is
used based on the respective grpc_address parameters.

LND requires a maximum fee to the `SendPaymentV2` call, so CLN now also
takes into account a maximum fee.

LND doesn't have out-of-the-box support for labels on payments. Labels
are used to associate payment requests inside swapd with the payments on
the node. This is to associate utxos used for the payment with the
actual payout. LND has its own internal table associating labels to the
payment index on the node.