itest: add integration test support for lnd
cln and lnd now have their own fixtures, custom made for swapd. cln
wraps the LightningNode object from pyln-testing, while lnd is a newly
built fixture/factory for the node. They both share the same interface
so tests can call functions on the node to execute actions.

There is a parameter `--node` added to the tests in conftest.py. lnd,
cln or both can be passed there. By default it tests both. If one of the
flags is set it only tests swapd with that specific node implementation.
The 'user' is always a core lightning node. Only the node attached to
swapd is switched based on context.

The bitcoind fixture had to be copied from pyln-testing, because it
needed zmqpubrawblock and zmqpubrawtx on startup.