lock and add payment attempt in a transaction
If the payment attempt and lock are not added in a transaction, the
payment monitor has to contain logic to distinguish when a payment was
locked but an attempt was not added. By putting these in a transaction,
the payment is always locked when a payment attempt is added, and always
unlocked when the payment result is added. Making it easier to determine
which payment attempts are stale.